refactor(expm): move *_auto* free functions out of lib.rs into expm_impl.rs
`lib.rs` is now just module declarations + re-exports; the four
`expm_multiply_*_auto*` wrappers live in their own `expm_impl` module.
